pyproject.toml
[tool.poetry]
name = "token-analysis"
version = "0.1.0"
description = "get data from stock and forecast it"
authors = ["Yura Gorlo <yuragorlo@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12"
crewai = "^0.11.0"
unstructured = '==0.10.25'
pyowm = '3.3.0'
tools = "^0.1.9"
wikipedia = "1.4.0"
yfinance = "0.2.35"
sec-api = "1.0.17"
tiktoken = "0.5.2"
faiss-cpu = "1.7.4"
python-dotenv = "1.0.0"
openai = "1.7.1"
pandas = "2.2.1"
requests = "2.31.0"
datasetsforecast = "0.0.8"
neuralforecast = "1.7.0"
termcolor = "2.4.0"
torch = "2.2.2"
torchmetrics = "1.3.2"
matplotlib = "3.8.3"
langchain = "^0.1.16"
langchain-community = "^0.0.32"
langchain-core = "^0.1.42"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
